requirements.txt
aioice==0.10.1
aiortc==1.13.0
altair==5.5.0
attrs==25.3.0
av==14.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
dnspython==2.7.0
filelock==3.18.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
huggingface-hub==0.32.6
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.42.0
networkx==3.5
numpy==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.3.0
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
safetensors==0.5.3
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
streamlit-webrtc==0.62.4
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
